- Reorganized by putting all PrimitiveByteArray tests into one class
