--- src/test/java/com/ansill/arrays/PrimitiveByteArrayTest.java
@@ -1,9 +1,0 @@
-package com.ansill.arrays;
-
-import org.junit.jupiter.api.DisplayName;
-import test.ByteArrayTest;
-
-@DisplayName("PrimitiveByteArray tests")
-public interface PrimitiveByteArrayTest extends ByteArrayTest{
-
-}
